Add DTO validations and exception handling + Swagger setup

+++ src/main/java/com/arturo/tasker/controller/TaskController.java
@@ -1,10 +1,6 @@
 package com.arturo.tasker.controller;
 
-import java.time.format.DateTimeFormatter;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,9 +18,9 @@ import com.arturo.tasker.dto.TaskResponse;
 import com.arturo.tasker.entity.Task;
 import com.arturo.tasker.entity.User;
 import com.arturo.tasker.service.TaskService;
-import com.arturo.tasker.service.UserService;
 
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -33,22 +29,12 @@ import lombok.RequiredArgsConstructor;
 public class TaskController {
 
 	private final TaskService taskService;
-	private final UserService userService;
-	
-	private final DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-	
+		
 	@Operation(summary = "Access to personal tasks")
 	@GetMapping
     public List<TaskResponse> getMyTasks() {
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		System.out.println("AUTH TYPE: " + auth.getClass());
-		System.out.println("PRINCIPAL: " + auth.getPrincipal());
-
-		if (auth == null || !auth.isAuthenticated() || "anonymousUser".equals(auth.getPrincipal())) {
-		    throw new RuntimeException("Not authenticated");
-		}
-
-		User currentUser = (User) auth.getPrincipal();
+		
+		User currentUser = getCurrentUser();
 		
         return taskService.getAllTasksByUser(currentUser.getId());
     }
@@ -56,14 +42,8 @@ public class TaskController {
 	@Operation(summary = "Access to a certain personal task")
 	@GetMapping("/{id}")
 	public ResponseEntity<TaskResponse> getMyTask(@PathVariable Long id) {
-		
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		
-			if (auth == null || !auth.isAuthenticated() || "anonymousUser".equals(auth.getPrincipal())) {
-				throw new RuntimeException("Not authenticated");
-			}
-						
-			User currentUser = (User) auth.getPrincipal();
+				
+			User currentUser = getCurrentUser();
 			
 			Task task = taskService.findById(id)
 					.orElseThrow(() -> new RuntimeException("Task not found"));
@@ -79,14 +59,23 @@ public class TaskController {
 			
 			}
 	
+	@Operation(summary = "Update one of your tasks")
 	@PutMapping("/{id}")
 	public ResponseEntity<TaskResponse> updateTask(
+			@Valid
 			@PathVariable Long id,
 			@RequestBody TaskRequest request
 		) {
-		Task existing = taskService.findById(id)
-				.orElseThrow(() -> new RuntimeException("Task not found"));
-		
+        User currentUser = getCurrentUser();
+
+        Task existing = taskService.findById(id)
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+
+        // The task must belong to the current user
+		if (!existing.getUser().getId().equals(currentUser.getId())) {
+            throw new RuntimeException("Forbidden");
+        }
+				
 		existing.setTitle(request.getTitle());
 		existing.setDescription(request.getDescription());
 		existing.setCompleted(request.isCompleted());
@@ -96,8 +85,17 @@ public class TaskController {
 		return ResponseEntity.ok(toResponse(updated));
 	}
 	
+	@Operation(summary = "Delete one of your tasks")
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteTask(@PathVariable Long id) {
+		User currentUser = getCurrentUser();
+		Task existing = taskService.findById(id)
+                .orElseThrow(() -> new RuntimeException("Task not found"));
+
+        // The task must belong to the current user
+		if (!existing.getUser().getId().equals(currentUser.getId())) {
+            throw new RuntimeException("Forbidden");
+        }
 		taskService.deleteById(id);
 		return ResponseEntity.noContent().build();
 	}
@@ -113,13 +111,22 @@ public class TaskController {
 				.build();
 	}
 	
+	@Operation(summary = "Create a task")
 	@PostMapping
-	public TaskResponse create(@RequestBody TaskRequest request) {
+	public TaskResponse create(@Valid @RequestBody TaskRequest request) {
 		
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-
-		User currentUser = (User) auth.getPrincipal();
+		User currentUser = getCurrentUser();
 		
 		return taskService.createTask(request, currentUser.getId());
 	}
+	
+	private User getCurrentUser() {
+	    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+	    if (auth == null || !auth.isAuthenticated() || "anonymousUser".equals(auth.getPrincipal())) {
+	        throw new RuntimeException("Not authenticated");
+	    }
+	    return (User) auth.getPrincipal();
+	}
+
 }
